Export search functionality into its own package
Package search can be implemented quite easily from within the `pkg/`
framework, and it's and endpoint which does not use any authentication
at all.

In order to test this, the package search functionality from
`internal/connect/` has been removed and everything for the zypper
plugin `search-packages` has been migrated to this new interface.
