+++ cmd/zypper-search-packages/search-packages.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/SUSE/connect-ng/internal/connect"
 	"github.com/SUSE/connect-ng/internal/util"
-	"github.com/SUSE/connect-ng/pkg/registration"
 )
 
 var (
@@ -163,10 +162,8 @@ func main() {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	// TODO(mssola): to be removed by the end of RR4.
-	connect.CFG = opts
 
-	results := searchInModules(flag.Args(), matchExact, caseSensitive)
+	results := searchInModules(opts, flag.Args(), matchExact, caseSensitive)
 	if !noLocalRepos {
 		repoResults := searchInRepos(flag.Args(), matchExact, caseSensitive)
 		results = append(results, repoResults...)
@@ -344,10 +341,10 @@ func searchInRepos(patterns []string, matchExact, caseSensitive bool) []searchRe
 	return ret
 }
 
-func searchInModules(patterns []string, matchExact, caseSensitive bool) []searchResult {
+func searchInModules(opts *connect.Options, patterns []string, matchExact, caseSensitive bool) []searchResult {
 	ret := make([]searchResult, 0)
 	for _, query := range patterns {
-		found, err := connect.SearchPackage(query, registration.Product{})
+		found, err := connect.SearchPackage(opts, query)
 		if err != nil {
 			fmt.Printf("Could not search for the package: %v", err)
 		}
@@ -361,9 +358,9 @@ func searchInModules(patterns []string, matchExact, caseSensitive bool) []search
 					Version:     pkg.Version,
 					Release:     pkg.Release,
 					Arch:        pkg.Arch,
-					ProdName:    fmt.Sprintf("%s (%s)", p.Name, p.Ident), // TODO: anything better? do we need raw p.Name?
-					ProdIdent:   p.Ident,
-					ProdEdition: p.Edition,
+					ProdName:    fmt.Sprintf("%s (%s)", p.Name, p.Identifier), // TODO: anything better? do we need raw p.Name?
+					ProdIdent:   p.Identifier,
+					ProdEdition: p.Edition(),
 					ProdArch:    p.Arch,
 					ProdFree:    p.Free,
 				}
